[bug 782292] Add KB Helpfulness chart to KB Dashboard.
