+++ apps/dashboards/templates/dashboards/contributors.html
@@ -3,7 +3,7 @@
 {% from "includes/common_macros.html" import for_contributors_sidebar %}
 {% from "dashboards/includes/macros.html" import print_readout, print_subscription_menu with context %}
 {% set title = _('Knowledge Base Dashboard') %}
-{% set scripts = ('wiki', 'dashboard.helpful', 'highcharts',) %}
+{% set scripts = ('wiki', 'highcharts', 'wiki.dashboard', ) %}
 {% set styles = ('kbdashboards',) %}
 {% set crumbs = [(None, title)] %}
 {# TODO: Move required dashboard scripts out of wiki.js into their own #}
@@ -41,6 +41,20 @@
       {% for readout in readouts.itervalues() %}
         {{ print_readout(readout, locale=default_locale) }}
       {% endfor %}
+
+      <div id="kb-helpfulness-chart"
+        data-vote-url="{{ url('api_dispatch_list', resource_name='kpi_kb_vote', api_name='v1') }}">
+      </div>
+      <!-- Backbone View Templates -->
+      <script type="text/template" id="chart-template">
+        <select class="grouping">
+          <option value="d">{{ _('Daily') }}</option>
+          <option value="w">{{ _('Weekly') }}</option>
+          <option value="m">{{ _('Monthly') }}</option>
+        </select>
+        <div class="placeholder"></div>
+      </script>
+
       <a id="show-chart">Show Aggregate Helpfulness Chart</a>
       <div id="helpful-chart-tall" data-url="{{ url('dashboards.get_helpful_graph_async') }}"></div>
     </article>
