+++ apps/kpi/urls.py
@@ -2,13 +2,14 @@ from django.conf.urls.defaults import patterns, url, include
 
 from kpi.api import (QuestionsResource, VoteResource,
                      ActiveContributorsResource, ElasticClickthroughResource,
-                     VisitorsResource, L10nCoverageResource)
+                     VisitorsResource, L10nCoverageResource, KBVoteResource)
 from tastypie.api import Api
 
 
 v1_api = Api(api_name='v1')
 v1_api.register(QuestionsResource())
 v1_api.register(VoteResource())
+v1_api.register(KBVoteResource())
 v1_api.register(ActiveContributorsResource())
 v1_api.register(ElasticClickthroughResource())
 v1_api.register(VisitorsResource())
