+++ media/less/kbdashboards.less
@@ -436,3 +436,30 @@
     }
   }
 }
+
+#kb-helpfulness-chart {
+  min-height: 400px;
+
+  section {
+    background: url('../img/wait-trans.gif') no-repeat center center;
+    height: 400px;
+    margin: 20px 0;
+    width: 680px;
+  }
+
+  .graph {
+    padding: 10px;
+    position: relative;
+  }
+
+  .grouping {
+    position: absolute;
+    right: 20px;
+    top: 50px;
+    z-index: 2;
+  }
+
+  .placeholder {
+    z-index: 1;
+  }
+}
