+++ settings.py
@@ -652,7 +652,10 @@ MINIFY_BUNDLES = {
             'js/libs/ace/src-min/ace.js',
             'js/libs/ace.mode-sumo.js',
         ),
-        'dashboard.helpful': (
+        'wiki.dashboard': (
+            'js/libs/backbone.js',
+            'js/charts.js',
+            'js/wiki.dashboard.js',
             'js/helpfuldashcharts.js',
         ),
         'highcharts': (
@@ -721,6 +724,7 @@ MINIFY_BUNDLES = {
         ),
         'kpi.dashboard': (
             'js/libs/backbone.js',
+            'js/charts.js',
             'js/kpi.dashboard.js',
         ),
         'products-mobile': (
